+++ protocols/mdns/tests/use-tokio.rs
@@ -71,33 +71,27 @@ async fn run_discovery_test(config: MdnsConfig) -> Result<(), Box<dyn Error>> {
     let mut discovered_b = false;
     loop {
         futures::select! {
-            ev = a.select_next_some() => match ev {
-                SwarmEvent::Behaviour(MdnsEvent::Discovered(peers)) => {
-                    for (peer, _addr) in peers {
-                        if peer == *b.local_peer_id() {
-                            if discovered_a {
-                                return Ok(());
-                            } else {
-                                discovered_b = true;
-                            }
+            ev = a.select_next_some() => if let SwarmEvent::Behaviour(MdnsEvent::Discovered(peers)) = ev {
+                for (peer, _addr) in peers {
+                    if peer == *b.local_peer_id() {
+                        if discovered_a {
+                            return Ok(());
+                        } else {
+                            discovered_b = true;
                         }
                     }
                 }
-                _ => {}
             },
-            ev = b.select_next_some() => match ev {
-                SwarmEvent::Behaviour(MdnsEvent::Discovered(peers)) => {
-                    for (peer, _addr) in peers {
-                        if peer == *a.local_peer_id() {
-                            if discovered_b {
-                                return Ok(());
-                            } else {
-                                discovered_a = true;
-                            }
+            ev = b.select_next_some() => if let SwarmEvent::Behaviour(MdnsEvent::Discovered(peers)) = ev {
+                for (peer, _addr) in peers {
+                    if peer == *a.local_peer_id() {
+                        if discovered_b {
+                            return Ok(());
+                        } else {
+                            discovered_a = true;
                         }
                     }
                 }
-                _ => {}
             }
         }
     }